+++ src/symmetric_crypto/key.rs
@@ -1,101 +1,83 @@
 //! Define a symmetric key object of variable size.
 
 use crate::{symmetric_crypto::SymKey, CryptoCoreError, KeyTrait};
-use aes::cipher::generic_array::{ArrayLength, GenericArray};
+use core::{convert::TryFrom, fmt::Display, hash::Hash, ops::Deref};
 use rand_core::{CryptoRng, RngCore};
-use serde::{Deserialize, Serialize};
-use std::{convert::TryFrom, fmt::Display, ops::Deref};
 use zeroize::{Zeroize, ZeroizeOnDrop};
 
 /// Symmetric key of a given size.
 ///
 /// It is internally built using an array of bytes of the given length.
-#[derive(Debug, Default, Hash, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Key<KeyLength: ArrayLength<u8>>(GenericArray<u8, KeyLength>);
+#[derive(Debug, Hash, Clone, PartialEq, Eq)]
+pub struct Key<const LENGTH: usize>([u8; LENGTH]);
 
-impl<KeyLength: ArrayLength<u8>> Key<KeyLength> {
-    /// Generate a new symmetric random `Key`
-    pub fn new<R: RngCore + CryptoRng>(rng: &mut R) -> Self {
-        let mut key = Self(GenericArray::<u8, KeyLength>::default());
-        rng.fill_bytes(&mut key.0);
-        key
-    }
-}
-
-impl<KeyLength: Eq + ArrayLength<u8>> KeyTrait for Key<KeyLength> {
-    type Length = KeyLength;
-
-    /// Convert the given key into bytes, with copy.
-    fn to_bytes(&self) -> generic_array::GenericArray<u8, KeyLength> {
-        self.0.clone()
+impl<const LENGTH: usize> KeyTrait<LENGTH> for Key<LENGTH> {
+    /// Generates a new symmetric random `Key`.
+    #[inline]
+    fn new<R: RngCore + CryptoRng>(rng: &mut R) -> Self {
+        let mut key = [0; LENGTH];
+        rng.fill_bytes(&mut key);
+        Self(key)
     }
 
-    /// Try to convert the given bytes into a key. Size must be correct.
+    /// Converts the given key into bytes.
+    #[inline]
+    fn to_bytes(&self) -> [u8; LENGTH] {
+        self.0.to_owned()
+    }
+
+    /// Tries to convert the given bytes into a key. Size must be correct.
+    #[inline]
     fn try_from_bytes(bytes: &[u8]) -> Result<Self, CryptoCoreError> {
-        Self::try_from(bytes)
+        let bytes = <[u8; LENGTH]>::try_from(bytes)
+            .map_err(|e| CryptoCoreError::ConversionError(e.to_string()))?;
+        Ok(Self(bytes))
     }
 }
 
-impl<KeyLength: Eq + ArrayLength<u8>> SymKey for Key<KeyLength> {
-    /// Convert the given key into a byte slice, without copy.
+impl<const LENGTH: usize> SymKey<LENGTH> for Key<LENGTH> {
+    /// Converts the given key into a byte slice.
+    #[inline]
     fn as_bytes(&self) -> &[u8] {
         &self.0
     }
 
-    fn from_bytes(bytes: GenericArray<u8, Self::Length>) -> Self {
+    /// Consumes the key to return underlying bytes.
+    #[inline]
+    fn into_bytes(self) -> [u8; LENGTH] {
+        self.0
+    }
+
+    /// Converts the given bytes with correct size into a key.
+    #[inline]
+    fn from_bytes(bytes: [u8; LENGTH]) -> Self {
         Self(bytes)
     }
 }
 
-impl<KeyLength: Eq + ArrayLength<u8>> From<Key<KeyLength>> for GenericArray<u8, KeyLength> {
-    fn from(k: Key<KeyLength>) -> Self {
-        k.to_bytes()
+impl<const KEY_LENGTH: usize> Display for Key<KEY_LENGTH> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{}", hex::encode(self.as_bytes()))
     }
 }
 
-impl<KeyLength: ArrayLength<u8>> From<GenericArray<u8, KeyLength>> for Key<KeyLength> {
-    fn from(b: GenericArray<u8, KeyLength>) -> Self {
-        Self(b)
-    }
-}
-
-impl<'a, KeyLength: ArrayLength<u8>> TryFrom<&'a [u8]> for Key<KeyLength> {
-    type Error = CryptoCoreError;
-
-    fn try_from(bytes: &'a [u8]) -> Result<Self, Self::Error> {
-        if bytes.len() != KeyLength::to_usize() {
-            return Err(Self::Error::SizeError {
-                given: bytes.len(),
-                expected: KeyLength::to_usize(),
-            });
-        }
-        Ok(Self(GenericArray::<u8, KeyLength>::clone_from_slice(bytes)))
-    }
-}
-
-impl<KeyLength: ArrayLength<u8>> Display for Key<KeyLength> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", hex::encode(self.as_slice()))
-    }
-}
-
-impl<KeyLength: ArrayLength<u8>> Zeroize for Key<KeyLength> {
+impl<const KEY_LENGTH: usize> Zeroize for Key<KEY_LENGTH> {
     fn zeroize(&mut self) {
         self.0.zeroize();
     }
 }
 
-// Implement `Drop` trait to follow R23.
-impl<KeyLength: ArrayLength<u8>> Drop for Key<KeyLength> {
+// Implements `Drop` trait to follow R23.
+impl<const KEY_LENGTH: usize> Drop for Key<KEY_LENGTH> {
     fn drop(&mut self) {
         self.zeroize();
     }
 }
 
-impl<KeyLength: ArrayLength<u8>> ZeroizeOnDrop for Key<KeyLength> {}
+impl<const KEY_LENGTH: usize> ZeroizeOnDrop for Key<KEY_LENGTH> {}
 
-impl<KeyLength: ArrayLength<u8>> Deref for Key<KeyLength> {
-    type Target = GenericArray<u8, KeyLength>;
+impl<const KEY_LENGTH: usize> Deref for Key<KEY_LENGTH> {
+    type Target = [u8];
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -106,24 +88,16 @@ impl<KeyLength: ArrayLength<u8>> Deref for Key<KeyLength> {
 mod tests {
 
     use crate::{entropy::CsRng, symmetric_crypto::key::Key, KeyTrait};
-    use generic_array::typenum::{ToInt, U32};
+
+    const KEY_LENGTH: usize = 32;
 
     #[test]
     fn test_key() {
         let mut cs_rng = CsRng::new();
-        let key_1 = Key::<U32>::new(&mut cs_rng);
-        assert_eq!(<U32 as ToInt<usize>>::to_int(), key_1.len());
+        let key_1 = Key::<KEY_LENGTH>::new(&mut cs_rng);
+        assert_eq!(KEY_LENGTH, key_1.len());
         let key_2 = Key::new(&mut cs_rng);
-        assert_eq!(<U32 as ToInt<usize>>::to_int(), key_2.len());
+        assert_eq!(KEY_LENGTH, key_2.len());
         assert_ne!(key_1, key_2);
-    }
-
-    #[test]
-    fn test_key_serialization() {
-        let mut cs_rng = CsRng::new();
-        let key = Key::<U32>::new(&mut cs_rng);
-        let bytes = key.to_bytes();
-        let res = Key::try_from(bytes).unwrap();
-        assert_eq!(key, res);
     }
 }