+++ src/symmetric_crypto/nonce.rs
@@ -1,41 +1,34 @@
-//! Define a nonce object, for use in symmetric encryption.
+//! Defines a nonce object, for use in symmetric encryption.
 //!
 //! A nonce, for Number used ONCE, is a randomly generated number used to
 //! ensure a ciphertext cannot be reused, hence avoiding replay attacks.
 
 use crate::CryptoCoreError;
-use num_bigint::BigUint;
-use rand_core::{CryptoRng, RngCore};
-use std::{
-    cmp::min,
+use core::{
     convert::{TryFrom, TryInto},
     fmt::{Debug, Display},
-    vec::Vec,
 };
+use rand_core::{CryptoRng, RngCore};
 
-/// Trait defining a nonce for use in a symmetric encryption scheme.
+/// Defines a nonce to use in a symmetric encryption scheme.
 pub trait NonceTrait: Send + Sync + Sized + Clone {
     /// Size of the nonce in bytes.
     const LENGTH: usize;
 
-    /// Generate a new nonce object.
+    /// Generates a new nonce object.
     #[must_use]
     fn new<R: RngCore + CryptoRng>(rng: &mut R) -> Self;
 
-    /// Try to deserialize the given `bytes` into a nonce object. The number of
-    /// `bytes` must be equal to `Self::LENGTH`.
+    /// Tries to deserialize the given `bytes` into a nonce object. The number
+    /// of `bytes` must be equal to `Self::LENGTH`.
     fn try_from_bytes(bytes: &[u8]) -> Result<Self, CryptoCoreError>;
-
-    /// Increment the nonce by the given value.
-    #[must_use]
-    fn increment(&self, increment: usize) -> Self;
 
     /// Xor the nonce with the given value.
     #[must_use]
     fn xor(&self, b2: &[u8]) -> Self;
 
-    /// Serialize the nonce.
-    fn as_slice(&self) -> &[u8];
+    /// Serializes the nonce.
+    fn as_bytes(&self) -> &[u8];
 }
 
 /// Nonce object of the given size.
@@ -47,12 +40,14 @@ pub struct Nonce<const NONCE_LENGTH: usize>([u8; NONCE_LENGTH]);
 impl<const NONCE_LENGTH: usize> NonceTrait for Nonce<NONCE_LENGTH> {
     const LENGTH: usize = NONCE_LENGTH;
 
+    #[inline]
     fn new<R: RngCore + CryptoRng>(rng: &mut R) -> Self {
-        let mut bytes = [0_u8; NONCE_LENGTH];
+        let mut bytes = [0; NONCE_LENGTH];
         rng.fill_bytes(&mut bytes);
         Self(bytes)
     }
 
+    #[inline]
     fn try_from_bytes(bytes: &[u8]) -> Result<Self, CryptoCoreError> {
         let b: [u8; NONCE_LENGTH] = bytes.try_into().map_err(|_| CryptoCoreError::SizeError {
             given: bytes.len(),
@@ -61,32 +56,18 @@ impl<const NONCE_LENGTH: usize> NonceTrait for Nonce<NONCE_LENGTH> {
         Ok(Self(b))
     }
 
-    fn increment(&self, increment: usize) -> Self {
-        let mut bi = BigUint::from_bytes_le(&self.0);
-        bi += BigUint::from(increment);
-        let mut bi_bytes = bi.to_bytes_le();
-        bi_bytes.resize(NONCE_LENGTH, 0);
-        Self(bi_bytes.try_into().expect("This should never happen"))
-    }
-
+    #[inline]
     fn xor(&self, b2: &[u8]) -> Self {
         let mut n = self.0;
-        for i in 0..min(b2.len(), NONCE_LENGTH) {
-            n[i] ^= b2[i];
+        for (ni, bi) in n.iter_mut().zip(b2) {
+            *ni ^= bi
         }
         Self(n)
     }
 
-    fn as_slice(&self) -> &[u8] {
+    #[inline]
+    fn as_bytes(&self) -> &[u8] {
         &self.0
-    }
-}
-
-impl<const NONCE_LENGTH: usize> TryFrom<Vec<u8>> for Nonce<NONCE_LENGTH> {
-    type Error = CryptoCoreError;
-
-    fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
-        Self::try_from_bytes(&bytes)
     }
 }
 
@@ -104,14 +85,8 @@ impl<const NONCE_LENGTH: usize> From<[u8; NONCE_LENGTH]> for Nonce<NONCE_LENGTH>
     }
 }
 
-impl<const NONCE_LENGTH: usize> From<Nonce<NONCE_LENGTH>> for Vec<u8> {
-    fn from(n: Nonce<NONCE_LENGTH>) -> Self {
-        n.0.to_vec()
-    }
-}
-
 impl<const NONCE_LENGTH: usize> Display for Nonce<NONCE_LENGTH> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{}", hex::encode(self.0))
     }
 }
@@ -123,25 +98,15 @@ mod tests {
         symmetric_crypto::nonce::{Nonce, NonceTrait},
     };
 
-    const NONCE_LENGTH: usize = 128;
+    const NONCE_LENGTH: usize = 12;
 
     #[test]
     fn test_nonce() {
         let mut cs_rng = CsRng::new();
         let nonce_1 = Nonce::<NONCE_LENGTH>::new(&mut cs_rng);
-        assert_eq!(NONCE_LENGTH, nonce_1.as_slice().len());
+        assert_eq!(NONCE_LENGTH, nonce_1.as_bytes().len());
         let nonce_2 = Nonce::<NONCE_LENGTH>::new(&mut cs_rng);
-        assert_eq!(NONCE_LENGTH, nonce_2.as_slice().len());
+        assert_eq!(NONCE_LENGTH, nonce_2.as_bytes().len());
         assert_ne!(nonce_1, nonce_2);
-    }
-
-    #[test]
-    fn test_increment_nonce() {
-        const NONCE_LENGTH: usize = 12;
-        let mut nonce: Nonce<NONCE_LENGTH> = Nonce::from([0_u8; NONCE_LENGTH]);
-        let inc = 1_usize << 10;
-        nonce = nonce.increment(inc);
-        println!("{}", hex::encode(nonce.0));
-        assert_eq!("000400000000000000000000", hex::encode(nonce.0));
     }
 }